Implement realtime update no need to refresh to reflect new data
